+++ module/product/tests/Application/Form/Product/Relation/ProductChildBySkusFormTest.php
@@ -8,6 +8,7 @@ declare(strict_types=1);
 
 namespace Ergonode\Product\Tests\Application\Form\Product\Relation;
 
+use Ergonode\Product\Domain\Entity\VariableProduct;
 use Symfony\Component\Form\Test\TypeTestCase;
 use Ergonode\Product\Application\Form\Product\Relation\ProductChildBySkusForm;
 use Ergonode\Product\Application\Model\Product\Relation\ProductChildBySkusFormModel;
@@ -16,14 +17,15 @@ class ProductChildBySkusFormTest extends TypeTestCase
 {
     public function testSubmitValidData(): void
     {
+        $parentProduct = $this->createMock(VariableProduct::class);
         $formData = [
             'skus' => ['78750d71-b5d4-4e5b-a9df-25ff619e2148'],
         ];
 
-        $object = new ProductChildBySkusFormModel();
+        $object = new ProductChildBySkusFormModel($parentProduct);
         $object->skus = ['78750d71-b5d4-4e5b-a9df-25ff619e2148'];
 
-        $objectToCompare = new ProductChildBySkusFormModel();
+        $objectToCompare = new ProductChildBySkusFormModel($parentProduct);
         $form = $this->factory->create(ProductChildBySkusForm::class, $objectToCompare);
         $form->submit($formData);